Add feedback about issues loading column data for presented options
When loading presented options for a dropdown type filter to allow the
CMS administrator to select valid values from a set, it is entirely
plausible that something such as a networking issue could occur and
prevent data from actually loading. Instead of leaving the user confused
with a never-ending loading indicator, we instead should show an alert
that something untoward has happened.

This notice also includes the prompt to try again.

Also included is the removal of a hard coded reference to a specific
form field for this specific application of the react component, as was
missed in the parent commit (5da31dd92edcac598b7fec95fb05ecd30af13499)
and general design tidies (such as presenting description text as other
form descriptions are throughout the CMS).